Rough refactoring separating dictionary classes from rmap.py and removing inline semantic validation,  moving it to crds.ceritify's validate_mapping call.
